added a logic for dailyEmission

+++ src/main/java/igor/kos/est/dto/request/FoodRequest.java
@@ -4,8 +4,10 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.PositiveOrZero;
 
+import java.math.BigDecimal;
+
 public record FoodRequest(
         @NotBlank String name,
-        @PositiveOrZero @NotNull Double emissionFactor,
+        @PositiveOrZero @NotNull BigDecimal emissionFactor,
         @NotBlank String category) {
 }
