+++ src/main/java/igor/kos/est/dto/request/TransportationRequest.java
@@ -4,5 +4,9 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.PositiveOrZero;
 
-public record TransportationRequest(@NotBlank String name, @NotNull @PositiveOrZero Double emissionFactor) {
+import java.math.BigDecimal;
+
+public record TransportationRequest(
+        @NotBlank String name,
+        @NotNull @PositiveOrZero BigDecimal emissionFactor) {
 }
