+++ src/main/java/igor/kos/est/dto/response/TransportationResponse.java
@@ -1,4 +1,6 @@
 package igor.kos.est.dto.response;
 
-public record TransportationResponse(Long id, String name, Double emissionFactor) {
+import java.math.BigDecimal;
+
+public record TransportationResponse(Long id, String name, BigDecimal emissionFactor) {
 }
