+++ src/main/java/igor/kos/est/entity/Utility.java
@@ -4,6 +4,8 @@ import igor.kos.est.enums.UtilityType;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @EqualsAndHashCode(callSuper = true)
 @Entity
 @Table(name = "utility")
@@ -14,7 +16,7 @@ import lombok.*;
 public class Utility extends BaseEntity {
 
     @Column(name = "emission_factor", nullable = false)
-    private Double emissionFactor;
+    private BigDecimal emissionFactor;
 
     @Enumerated(EnumType.STRING)
     @Column(name = "type", nullable = false, unique = true)
